+++ src/tracer/state.rs
@@ -155,20 +155,7 @@ impl<S: ReturnStack, I: Info + Clone + Default> State<S, I> {
                     self.stop_condition = StopCondition::Fused;
                     None
                 }
-                StopCondition::Sync {
-                    context,
-                    action: SyncAction::Compare,
-                } if hit_address_and_branch
-                    && context.privilege == self.privilege
-                    && self.last_insn.is_return_from_trap() =>
-                {
-                    self.stop_condition = StopCondition::Fused;
-                    None
-                }
-                StopCondition::Sync {
-                    context,
-                    action: SyncAction::Update,
-                } if hit_address_and_branch => {
+                StopCondition::Sync { context } if hit_address_and_branch => {
                     self.privilege = context.privilege;
                     self.stop_condition = StopCondition::Fused;
                     Some(context)
@@ -482,23 +469,9 @@ pub enum StopCondition {
     /// Stop when reaching a condition provided by an address packet
     Address { notify: bool, not_updiscon: bool },
     /// Stop at synchonization point (defined in sync packet)
-    Sync {
-        /// Context given in the sync packet
-        context: Context,
-        /// Action to perform at the synchronization point
-        action: SyncAction,
-    },
+    Sync { context: Context },
     /// The state is already fused and shall not be advanced
     Fused,
-}
-
-/// Synchronization action
-#[derive(Copy, Clone, Debug, PartialEq, Eq)]
-pub enum SyncAction {
-    /// Stop only if the sync context matches the [`State`]'s context
-    Compare,
-    /// Update the [`State`]'s context from the sync context
-    Update,
 }
 
 impl Default for StopCondition {